move getSlotContainerStates to utils/cluster
